Beginnings of the damage dealing by button click systems
There are lots of little things to clean up. First, the buttons associated with the enemies have no way of tracking with the enemy when it moves on screen.  Second, focus issues are preventing enemy target buttons from being clicked, because the virtual joystick keeps stealing the click away.  Other than that, the logic of the system is sound and just needs all the cool features put into it